added JPA specification to filter

+++ src/main/java/finance_management/service/BudgetService.java
@@ -3,13 +3,14 @@ package finance_management.service;
 import finance_management.dto.budget.BudgetRequest;
 import finance_management.dto.budget.BudgetResponse;
 import finance_management.exceptions.BudgetNotFoundException;
+import finance_management.exceptions.IncorrectUserException;
 import finance_management.exceptions.UserNotFoundException;
 import finance_management.mapper.BudgetMapper;
 import finance_management.model.Budget;
 import finance_management.model.User;
 import finance_management.repo.BudgetRepo;
 import finance_management.repo.UserRepo;
-import finance_management.validations.AmountValidation;
+import finance_management.validations.AmountAndFieldValidation;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -22,20 +23,20 @@ import java.util.Objects;
 public class BudgetService {
     private final BudgetRepo budgetRepo;
     private final UserRepo userRepo;
-    private final AmountValidation amountValidation;
+    private final AmountAndFieldValidation amountAndFieldValidation;
 
     public BudgetService(BudgetRepo budgetRepo,
                          UserRepo userRepo,
-                         AmountValidation amountValidation){
+                         AmountAndFieldValidation amountAndFieldValidation){
         this.budgetRepo = budgetRepo;
         this.userRepo = userRepo;
-        this.amountValidation = amountValidation;
+        this.amountAndFieldValidation = amountAndFieldValidation;
     }
 
     public BudgetResponse addBudget(BudgetRequest budgetRequest){
         String userEmail = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
         User user = userRepo.findByEmail(userEmail).orElseThrow(()->new UserNotFoundException("In valid user ID", HttpStatus.NOT_FOUND));
-        amountValidation.validate(budgetRequest);
+        amountAndFieldValidation.validate(budgetRequest);
         Budget budget = BudgetMapper.toEntity(budgetRequest);
         budget.setUser(user);
         budget =  budgetRepo.save(budget);
@@ -53,15 +54,27 @@ public class BudgetService {
 
     @Transactional
     public BudgetResponse updateBudget(Long id, BudgetRequest budgetRequest){
-        Budget bd = budgetRepo.findById(id).orElseThrow(()->new BudgetNotFoundException("Budget Id not valid", HttpStatus.NOT_FOUND));
-        bd.updateBudget(budgetRequest);
-        return BudgetMapper.toResponse(bd);
+        String userEmail = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
+        Budget budget = budgetRepo.findById(id).orElseThrow(()->new BudgetNotFoundException("Budget Id not valid", HttpStatus.NOT_FOUND));
+        if(!budget.getUser().getEmail().equals(userEmail)) throw new IncorrectUserException("Unauthorized resource");
+        budget.updateBudget(budgetRequest);
+        return BudgetMapper.toResponse(budget);
     }
 
     public String deleteBudget(Long id){
+        String userEmail = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
         Budget budget = budgetRepo.findById(id).orElseThrow(()->new BudgetNotFoundException("Budget Id not valid", HttpStatus.NOT_FOUND));
+        if(!budget.getUser().getEmail().equals(userEmail)) throw new IncorrectUserException("Unauthorized resource");
         budgetRepo.delete(budget);
         return "deleted successfullt";
+    }
+
+    @Transactional
+    public String bulkAddBudget(List<BudgetRequest> budgetRequests){
+        for(BudgetRequest budgetRequest : budgetRequests){
+            addBudget(budgetRequest);
+        }
+        return "Budgets added successfully";
     }
 
 }
